% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_accuracy.R
\name{class_accuracy}
\alias{class_accuracy}
\title{Class accuracy}
\usage{
class_accuracy(df, truth, pred, pos_flag = 1)
}
\arguments{
\item{df}{tibble

  A data frame containing the truth and prediction columns}

\item{truth}{factor

  A column identifier for the true class results (that is a factor).
  This should be an unquoted column name.}

\item{pred}{factor

  A column identifier for the predicted class results (that is also a factor).
  This should be an unquoted column name.}

\item{pos_flag}{numeric/factor

 A numeric or string indicating the "positive" class.}
}
\value{
A tibble
}
\description{
`class_accuracy()` provides common classification metrics to convey results
business users will understand.
}
