% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscper.R
\name{fiscper_month}
\alias{fiscper_month}
\alias{fiscper_date}
\alias{fiscper_quarter}
\alias{fiscper_semester}
\alias{fiscper_workdays}
\alias{make_fiscper}
\title{Fiscper functions}
\usage{
fiscper_month(fiscper)

fiscper_date(fiscper)

fiscper_quarter(fiscper)

fiscper_semester(fiscper)

fiscper_workdays(fiscper)

make_fiscper(year, month)
}
\arguments{
\item{fiscper}{Column containing the fiscper}
}
\description{
Extract info from fiscper
}
\examples{
# Extract info from fiscper
your_df \%>\%
  mutate(year = fiscper_year(fiscper),
         month = fiscper_month(fiscper),
         quarter = fiscper_quarter(fiscper),
         semester = fiscper_semester(fiscper),
         work_days = fiscper_workdays(fiscper),
         date = fiscper_date(fiscper))

# Create fiscper from year & month columns
your_df \%>\%
  mutate(make_fiscper(year, month))
}
