% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscper.R
\name{fiscper_year}
\alias{fiscper_year}
\title{Fiscper functions}
\usage{
fiscper_year(fiscper)
}
\arguments{
\item{fiscper}{Column containing the fiscper}
}
\description{
Extract info from fiscper
}
\examples{
# Extract info from fiscper
your_df \%>\%
  mutate(year = fiscper_year(fiscper),
         month = fiscper_month(fiscper),
         quarter = fiscper_quarter(fiscper),
         semester = fiscper_semester(fiscper),
         work_days = fiscper_workdays(fiscper),
         date = fiscper_date(fiscper))

# Create fiscper from year & month columns
your_df \%>\%
  mutate(make_fiscper(year, month))
}
